+++ lib/src/mixins/mixins.dart
@@ -1,0 +1,6 @@
+library;
+
+import 'package:flutter/material.dart';
+import 'package:vxstate/vxstate.dart';
+
+part 'src/vx_state.dart';