+++ lib/src/mixins/src/vx_state.dart
@@ -1,0 +1,7 @@
+part of '../mixins.dart';
+
+mixin VxStateUtilsMixin<T extends StatefulWidget> on State<T> {
+  void genericSetState(BuildContext context, VxMutation mutation, {VxStatus? status}) {
+    if (status == VxStatus.success) WidgetsBinding.instance.addPostFrameCallback((_) => setState(() {}));
+  }
+}